locking down mutations with user authentication

+++ server/src/resolvers/Mutation.js
@@ -1,5 +1,6 @@
 import bcrypt from 'bcryptjs'
 import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
  
 const Mutation = {
     async createUser(parent, { data }, { prisma }, info) {
@@ -37,49 +38,75 @@ const Mutation = {
             token: jwt.sign({userId: user.id}, 'secret')
         }
     },
-    deleteUser(parent, { id }, { prisma }, info) {
+    deleteUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
         return prisma.mutation.deleteUser({
-            where: { id }
+            where: { id: userId }
         }, info)
     },
-    updateUser(parent, { id, data }, { prisma }, info) {
+    updateUser(parent, { data }, { prisma, request}, info) {
+        const userId = getUserId(request)
         return prisma.mutation.updateUser({
             where: {
-                id
+                id: userId
             },
             data
         }, info)
     },
-    createCity(parent, { data }, { prisma }, info) {
+    createCity(parent, { data }, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.createCity({
             data: {
                 name: data.name,
                 available: data.available,
                 major: {
                     connect: {
-                        id: data.major
+                        id: userId
                     }
                 }
             }
         }, info)
     },
-    deleteCity(parent, { id }, { prisma }, info) {
+    async deleteCity(parent, { id }, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const cityExists = await prisma.exists.City({
+            id,
+            major: {
+                id: userId
+            }
+        })
+        if (!cityExists) {
+            throw new Error("Not possible to delete City")
+        }
         return prisma.mutation.deleteCity({
             where: { id }
         }, info)
     },
-    updateCity(parent, { id, data }, { prisma }, info) {
+    async updateCity(parent, { id, data }, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const cityExists = await prisma.exists.City({
+            id,
+            major: {
+                id: userId
+            }
+        })
+        if(!cityExists) {
+            throw new Error("Not possible to update City")
+        }
         return prisma.mutation.updateCity({
             where: { id },
             data
         }, info)
     },
-    createVote(parent, { data }, { prisma }, info) {
+    async createVote(parent, { data }, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.createVote({
             data: {
                 author: {
                     connect: {
-                        id: data.author
+                        id: userId
                     }
                 },
                 city: {
@@ -90,15 +117,19 @@ const Mutation = {
             }
         }, info)
     },
-    deleteVote(parent, { id }, { prisma }, info) {
+    async deleteVote(parent, { id }, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const voteExists = await prisma.exists.Vote({
+            id,
+            author: {
+                id: userId
+            }
+        })
+        if(!voteExists) {
+            throw new Error("Not possible to delete vote")
+        }
         return prisma.mutation.deleteVote({
             where: { id }
-        }, info)
-    },
-    updateVote(parent, { id, data  }, { prisma }, info) {
-        return prisma.mutation.updateVote({
-            where: { id },
-            data
         }, info)
     }
 }
